requirements.txt
asgiref==3.8.1
attrs==23.2.0
black==24.3.0
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
Django==5.0.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.0
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.34.0
rpds-py==0.18.0
sqlparse==0.4.4
tzdata==2024.1
uritemplate==4.1.1
drf_spectacular_sidecar==2024.3.4